+++ src/index.html
@@ -11,5 +11,6 @@
   <div id="navbar-here"></div>
   <div id="auth"></div>
   <div id="component-here"></div>
+  <div id="event-div"></div>
 </body>
 </html>
